Cleanup scripting refactor

+++ test.sh
@@ -1,7 +1,16 @@
 #!/bin/bash
 random_id=$(cat /dev/urandom | tr -dc 'a-z0-9' | fold -w 10 | head -n 1)
 set -e -x
-# Create cluster
+# Build and upload our framework
+FRAMEWORK=$1
+export UNIVERSE_URL_PATH=frameworks/$FRAMEWORK/$FRAMEWORK-universe-url
+frameworks/$FRAMEWORK/./build.sh aws
+if [ ! -f "$UNIVERSE_URL_PATH" ]; then
+    echo "Missing universe URL file: $UNIVERSE_URL_PATH"
+    exit 1
+fi
+
+# Create out test cluster
 cat <<EOF > config.yaml
 launch_config_version: 1
 deployment_name: dcos-ci-test-infinity-$random_id
@@ -16,21 +25,11 @@ template_parameters:
 ssh_user: core
 EOF
 
-# launch our test cluster and let it provision in the background while we build
-# dcos-launch create
-
-# Build and upload our framework
-FRAMEWORK=$1
-export UNIVERSE_URL_PATH=/dcos-commons/frameworks/$FRAMEWORK/$FRAMEWORK-universe-url
-/dcos-commons/frameworks/$FRAMEWORK/./build.sh aws
-if [ ! -f "$UNIVERSE_URL_PATH" ]; then
-    echo "Missing universe URL file: $UNIVERSE_URL_PATH"
-    exit 1
-fi
-
-# Wait for our cluster to finish
+dcos-launch create
 dcos-launch wait
-# Setup the SSH key for shakedown to use
+
+# Setup the SSH key for shakedown to use. This is the only way to configure
+# shakedown to use this ssh key without using the shakdedown CLI
 mkdir -p ~/.ssh/
 cat cluster_info.json | jq -r .ssh_private_key > ~/.ssh/id_rsa
 chmod 600 ~/.ssh/id_rsa
@@ -39,11 +38,11 @@ chmod 600 ~/.ssh/id_rsa
 CLUSTER_URL=http://`dcos-launch describe | jq -r .masters[0].public_ip`
 dcos config set core.dcos_url $CLUSTER_URL
 dcos config set core.ssl_verify false
-/dcos-commons/tools/./dcos_login.py
+tools/./dcos_login.py
 for url in `cat $UNIVERSE_URL_PATH`; do
     dcos package repo add --index=0 `echo $url | cut -d / -f 5` $url
 done
 
 set +e
-py.test --teamcity -vv -s -m "sanity and not azure" /dcos-commons/frameworks/$FRAMEWORK/tests
+py.test --teamcity -m "sanity" frameworks/$FRAMEWORK/tests
 dcos-launch delete
